[Cinema:2/4] Purchase endpoint and custom error messages implemented
Now the user can purchase tickets. If the seat is already bought, he'll see an error message. Additionally, if it's an invalid seat, there'll be a different error message.

This project has only what is necessary pass the test. The complete version (with tests and everything else) is on the main folder.