[Cinema:3/4] Purchase and return token
Now the user purchases a ticket through the /POST purchase endpoint and receives a token. With this token, he's able to request a refund through the /POST return endpoint. After this, the seat will be available again.

This project has only what is necessary pass the test. The complete version (with tests and everything else) is on the main folder.